refactors actions into separate module
